better handling of orphaned thread situation
This is where we can't find the root of the thread to which a message
belongs.

Rather than treating it as not in a thread, simply create a fake
thread root for it. That way, if this happens to more than one message
in the same thread, they are still grouped together.