remote tkinter GUI
It's now just a simple command line script.

The expectation (for now) is that you run it once, on a single file,
and use the bot to import that into a single channel. That causes the
bot to then just wait indefinitely, and you need to Ctrl-C and re-run
the script if you want to do another import.

This is not the best model, and will change in the future.

Token and filename are passed as raw command line args (no
options). This will be improved in the future with real arg parsing.