add a verbose option
sort of, for now it's hardcoded
eventually will parse args

with verbose, we:
* include the full pretty print of parsed_messages (previously commented out)
* log all of the messages to be posted (previously always present)

I still have no clue how the slackord() method is getting access to
variables defined in other scopes (now parsed_messages and
verbose). But I'm not going to worry about it much, b/c I still plan
to change the overall structure to not use a bot.